Authorizer 세분화
`Authorizer`에 `Account`를 입력해야 하는것이 아니라, `Authorizer`가 `Account`를 찾도록 `Authorizer`클래스로써 기능을 확대하는 방향이 SOLID의 단일책임 원칙에 좀 더 부합하는 방향이라 판단

+++ src/main/java/kr/reciptopia/reciptopiaserver/business/service/CommentService.java
@@ -1,7 +1,7 @@
 package kr.reciptopia.reciptopiaserver.business.service;
 
 import java.util.List;
-import kr.reciptopia.reciptopiaserver.business.service.authorizer.AbstractAuthorizer;
+import kr.reciptopia.reciptopiaserver.business.service.authorizer.CommentAuthorizer;
 import kr.reciptopia.reciptopiaserver.business.service.helper.RepositoryHelper;
 import kr.reciptopia.reciptopiaserver.domain.dto.CommentDto.Create;
 import kr.reciptopia.reciptopiaserver.domain.dto.CommentDto.Result;
@@ -25,14 +25,14 @@ public class CommentService {
 
     private final CommentRepository commentRepository;
     private final RepositoryHelper repoHelper;
-    private final AbstractAuthorizer authorizer;
+    private final CommentAuthorizer commentAuthorizer;
 
     @Transactional
     public Result create(Create dto, Authentication authentication) {
         Comment comment = dto.asEntity();
 
         Account owner = repoHelper.findAccountOrThrow(dto.ownerId());
-        authorizer.requireByOneself(authentication, owner);
+        commentAuthorizer.requireByOneself(authentication, owner);
         comment.setOwner(owner);
 
         Post post = repoHelper.findPostOrThrow(dto.postId());
@@ -52,20 +52,20 @@ public class CommentService {
 
     @Transactional
     public Result update(Long id, Update dto, Authentication authentication) {
-        Comment entity = repoHelper.findCommentOrThrow(id);
-        authorizer.requireByOneself(authentication, entity.getOwner());
+        Comment comment = repoHelper.findCommentOrThrow(id);
+        commentAuthorizer.requireCommentOwner(authentication, comment);
 
         if (dto.content() != null) {
-            entity.setContent(dto.content());
+            comment.setContent(dto.content());
         }
 
-        return Result.of(commentRepository.save(entity));
+        return Result.of(commentRepository.save(comment));
     }
 
     @Transactional
     public void delete(Long id, Authentication authentication) {
         Comment comment = repoHelper.findCommentOrThrow(id);
-        authorizer.requireByOneself(authentication, comment.getOwner());
+        commentAuthorizer.requireCommentOwner(authentication, comment);
         comment.removeAllCollections();
 
         commentRepository.delete(comment);
